agentflow_horizon_backend/requirements.txt
fastapi
uvicorn
chromadb
google-generativeai
google-adk
python-dotenv
beautifulsoup4
requests
pypdf2
python-multipart
sqlalchemy
